REFACTOR : create-room event 레포지토리 주입하여 수정 #58
Changes to be committed:
	renamed:    src/games/dto/room.data.dto.ts -> src/games/dto/room.data.insert.dto.ts
	modified:   src/games/dto/roomInfoToMain.dto.ts
	modified:   src/games/dto/roomInfoToRoom.dto.ts
	modified:   src/games/dto/socketId-map.request.dto.ts
	modified:   src/games/entities/room.entity.ts
	modified:   src/games/entities/socketIdMap.entity.ts
	modified:   src/games/games.gateway.ts
	modified:   src/games/players.service.ts
	modified:   src/games/room.service.ts
	modified:   src/users/entities/token-map.entity.ts
	modified:   src/users/users.controller.ts
	modified:   src/users/users.service.ts

+++ project/src/games/entities/room.entity.ts
@@ -34,7 +34,7 @@ export class Room {
     @Column()
     isSecreteRoom: boolean;
 
-    @Column({ type: 'tinyint', width: 4 })
+    @Column({ type: 'int', width: 4 })
     roomPassword: number;
 
     @Column()
